Time: 492 ms (18.79%), Space: 409.4 MB (10.74%) - LeetHub

+++ 3558-number-of-ways-to-assign-edge-weights-i/3558-number-of-ways-to-assign-edge-weights-i.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long assignEdgeWeights(vector<vector<int>>& edges) {
+    int assignEdgeWeights(vector<vector<int>>& edges) {
         unordered_set<int> us;
         unordered_map<int, vector<int>> um;
         long long MOD = 1e09 + 7;
